Thats are the files of lib folder of your proyect
For this files are required fl_chart and scidart packages. The main file manages the structure of the app, plot is a widget that makes the line chart and difussion.dart makes the implicit and explicit methods and return a list of values that are inserted like parameter in plot function.